+++ test/views/users/store.conf.js
@@ -1,4 +1,5 @@
 import app from '@/store/modules/app'
+import settings from '@/store/modules/settings'
 import user from '@/store/modules/user'
 import userProfile from '@/store/modules/userProfile'
 import users from '@/store/modules/users'
@@ -7,6 +8,7 @@ import getters from '@/store/getters'
 export default {
   modules: {
     app,
+    settings,
     user,
     userProfile,
     users